allow palette in 64 colour modes to be adjusted
the palette in 64 colour modes was fixed, but that palette did not match the logical colours, and could not be redefined, which was kinda bonkers

this changes the palette handling for 64 colour modes to allow the palette to be redefined

NB “logical” colours on Agon VDP are in bit format RRGGBB, i.e. with red in the most significant bits.  this is the opposite way around to video circuitry, and hence also the order of bits in RGBA2222 format