keep track of logical colour values
this allows the user chosen colours in 64 colour modes to keep track of palette changes, i.e. if text is in colour 1, and then the palette entry for colour 1 is changed then subsequent prints will use the new colour

previously colours selected in 64 colour modes would just “stick”, i.e. if text colour 1 was selected, which equated to physical colour 27, and then the palette was changed so that it then used physical colour 42 then text output would still be using physical colour 27.  this commit resolves that, so from the user’s perspective as they’ve selected (logical) colour 1 they will continue to be outputting in logical colour 1