fix strange intermittent crash
it seems that the a `>= 0` test when `-O3` is set, as used to check whether a byte has been read or not, will not _always_ be reliable, whereas a `!= 0` test will.  this commit changes tests to `!= 0` and also changes the compiler flag to `-O2` which also fixes the issue (a belt & braces style double fix).  testing shows essentially identical performance with `-O2`

also restore `timeout` argument to `discardBytes`