add “expand bitmap” buffer command
allows for bitmapped data where a pixel is of 1-7 bits in length to be expanded into bytes, using a given map.  the map for expansion can either be provided with the command, or as a pointer to a buffer.  a pixel width optionally can be provided to indicate at what pixel width a byte alignment should occur, if required