change our fixed-point format
following discussions with @calc84maniac, this changes our fixed-point format

the new format essentially has our range of values _start_ at being +/- 2^shift - so a shift value of 0 equates to -1 to +1.  the “binary point” essentially sits just to the right of the top-most bit.  shift is now a positive integer only and will move the “binary point” to the right in the value.

as the shift value is now in the range of 0-31, this lets 16-bit fixed-point numbers express the same range of values as 32-bit fixed point, just with less precision