allow flags to be used in buffered commands
extends “conditional” buffer commands to be able to check against flag values.  as part of this, support for checking 16-bit values has been added via setting the 0x80 bit of the conditional operation.  when the “use flag” bit (0x40) is set on the conditional operation an offset will not be read for the condition source value

adds a “read flag into buffer” command which will copy the value of a flag into a buffer.  this takes an options byte which is similar to conditional options, so a “use 16-bit” flag (0x80) is supported.  a default value can be set if the flag 0x40 is provided, which must be either 8 or 16-bits depending on the 16-bit flag