add support for temporary paged mode
temporary paged mode is enabled using `VDU 23,0,&9A` and remains active until there’s been no pending VDU commands for a frame, whereupon the actual paged mode setting will be restored

fixed an issue whereby holding ctrl would cause characters to go missing from output.  the root cause of this is that paged mode prevents the cursor down action when it engages - which can result in characters get drawn off the right edge of the screen.  the fix is to ensure that the `plotString` routine limits itself to the width of the screen