mouse management changes
moves mouse variables into the “system” number range to sit alongside keyboard variables

setting a cursor via variables allows you to make a cursor visible even when the mouse is enabled (or cannot be enabled)

when a bitmap is deleted, any mouse cursor defined that used it will now be cleared.  if it was the current cursor, the cursor will be reset to the default.  (this occurs whether the cursor was visible or not, but if the cursor was already hidden the default cursor will not get shown).  this was previously kinda buggy and could leave the mouse cursor using a deleted bitmap

interim commit - need to more properly solve mouse visibility control