keyboard variables and event
Adds variables for keyboard state, and the last keyboard event

also adds a keyboard event which will be fired by the main loop’s keyboard handler for every keyboard event it processes.

also included is support for reading variables in big-endian order.  this will switch the byte ordering of the variable when it is read.  this can allow only the upper byte in a variable to be read.  (NB this is not supported by conditions, only the read variable into buffer command)