add keymap variables
adds a block of 259 variables starting from 0x0800 which show whether a key is pressed, and it’s current ASCII value.  the index into this block is the fab-gl keycode

the lower byte of the variable will be 0/1 to indicate whether the key is pressed.  the upper byte is the ASCII conversion of the keycode.  the fab-gl function that returns the ASCII conversion applies modifiers meaning that it’s value may be altered by CTRL being pressed, scroll-lock will return ASCII_XON/XOFF, and it appears that Tab will only return a character (9) if the shift key is not being held, otherwise it will not be recognised and return a 0